requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore==2.11.0
firebase-admin==6.5.0
python-dotenv
pytrends
